# Share ChildBinding objects between siblings