It's hard to measure the benefit from this, but I used the following little snippet of Chisel to make a bundle of Bundles:
```scala
import chisel3._
// _root_ disambiguates from package chisel3.util.circt if user imports chisel3.util._
import _root_.circt.stage.ChiselStage

class MyBundle extends Bundle {
  val a, b, c, d, e, f, g = UInt(8.W)
}

class Foo(n: Int) extends Module {
  val in = IO(Input(Vec(n, new MyBundle)))
  val out = IO(Output(Vec(n, new MyBundle)))

  out :#= in
}

object Main extends App {
  val n = args(0).toInt
  ChiselStage
    .emitCHIRRTLFile(
      gen = new Foo(n)
    )
    .toSeq
}
```
Then I built a fat assembly jar with Mill, and checked the heap size with `find_heap_bound`:

```
./firrtl/benchmark/scripts/find_heap_bound.py -vvv --start-size 4G --min-step 100M --context 2 -- -cp assembly.jar Main 200000
```

With this, I measured a ~2.5% memory reduction for this very Bundle-heavy code.

I had thought about a more aggressive change, where `ChildBinding` becomes a `case object` and we instead get the `parent` from `ref` (which does duplicate this information). However, using the above benchmark, I measured the exact same memory improvement. Since I cannot measure any benefit from the much more aggressive and hacky alternative approach, I think we should do the simpler thing which gets more-or-less the same benefit.

### Contributor Checklist

- [x] Did you add Scaladoc to every public function/method?
- [ ] Did you add at least one test demonstrating the PR?
- [x] Did you delete any extraneous printlns/debugging code?
- [x] Did you specify the type of improvement?
- [ ] Did you add appropriate documentation in `docs/src`?
- [x] Did you request a desired merge strategy?
- [x] Did you add text to be included in the Release Notes for this change?

<!--
If you PR has any impact on the user API or affects backend code generation,
please describe the change in the "Release Notes" section below.
-->

#### Type of Improvement

- Performance improvement


#### Desired Merge Strategy

<!-- If approved, how should this PR be merged? Delete those that do not apply -->
- Squash

#### Release Notes

This reduces memory use by `n - 1` times 16-bytes for an Aggregate with `n` elements.

### Reviewer Checklist (only modified by reviewer)
- [ ] Did you add the appropriate labels? (Select the most appropriate one based on the "Type of Improvement")
- [ ] Did you mark the proper milestone (Bug fix: `3.6.x`, `5.x`, or `6.x` depending on impact, API modification or big change: `7.0`)?
- [ ] Did you review?
- [ ] Did you check whether all relevant Contributor checkboxes have been checked?
- [ ] Did you do one of the following when ready to merge:
  - [ ] Squash: You/ the contributor `Enable auto-merge (squash)`, clean up the commit message, and label with `Please Merge`.
  - [ ] Merge: Ensure that contributor has cleaned up their commit history, then merge with `Create a merge commit`.