@jackkoenig 2024-07-02
I validated that the ChildBinding objects are **not** currently shared and that with this change they are now shared with the Eclipse MemoryAnalyzer (https://eclipse.dev/mat/). This changed the measured memory use per UInt child of the Bundles in my example above from 72 bytes shallow, 144 retained to 72 bytes shallow, 128 retained, which measures my expectation. I'm also impressed that MAT is able to figure out that the individual elements should not be dinged for retaining the object that they all share.