@jackkoenig 2024-07-02
I figured out how to benchmark a little bit better by avoiding serialization

```scala
import chisel3._
// _root_ disambiguates from package chisel3.util.circt if user imports chisel3.util._
import _root_.circt.stage.ChiselStage

class MyBundle extends Bundle {
  val a, b, c, d, e, f, g = UInt(8.W)
}

class Foo(n: Int) extends Module {
  val in = IO(Input(Vec(n, new MyBundle)))
  val out = IO(Output(Vec(n, new MyBundle)))

  out :#= in
}

object Main extends App {
  val n = args(0).toInt
  val phase = new chisel3.stage.phases.Elaborate
  val annos = Seq(
    chisel3.stage.ChiselGeneratorAnnotation(() => new Foo(n))
  )
  println(phase.transform(annos).size)
}
```

Using the same basic approach but small min-step and more context:
```
./firrtl/benchmark/scripts/find_heap_bound.py -vvv --start-size 4G --min-step 10M --context 5 -- -cp assembly.jar Main 200000
```

Without this change:

Xmx | Max RSS (MiB) | Wall Clock (s) | User Time (s)
-- | -- | -- | --
1G | 1344 | 7.09 | 56.45
920M | 1269 | 6.92 | 88.26
910M | 1220 | 7.61 | 90.95
900M | 1198 | 9.98 | 116.99
890M | - | - | -
870M | - | - | -

With this change:

Xmx | Max RSS (MiB) | Wall Clock (s) | User Time (s)
-- | -- | -- | --
1G | 1324 | 6.75 | 74.59
880M | 1202 | 6.86 | 58.98
870M | 1221 | 7.1 | 82.23
860M | 1164 | 9.19 | 98.01
850M | - | - | -
840M | - | - | -

Using the best Max RSS for each shows a memory reduction of 2.8%. Small, but measurable!